+++ src/lib.rs
@@ -2,7 +2,7 @@
 //!
 //! This driver was built using [`embedded-hal`] traits.
 //!
-//! [`embedded-hal`]: https://docs.rs/embedded-hal/0.2
+//! [`embedded-hal`]: https://docs.rs/embedded-hal/1.0
 //!
 //! # Examples
 //!
@@ -14,39 +14,39 @@
 #![deny(warnings)]
 #![no_std]
 
+use embedded_hal::spi::Mode;
+use embedded_hal::spi::SpiDevice;
 
+/// Minimal time in nanoseconds between chip select assertion and clock edge.
+pub const MINIMUM_CS_SETUP_TIME_NS: u32 = 5;
 
-use embedded_hal::blocking::spi::{Transfer, Write};
-use embedded_hal::digital::v2::OutputPin;
-use embedded_hal::spi::{Mode};
+/// Expected WHO_AM_I register value for the L3GD20 sensor.
+pub const WHO_AM_I_L3GD20: u8 = 0xD4;
+/// Expected WHO_AM_I register value for the L3GD20H sensor.
+pub const WHO_AM_I_L3GD20H: u8 = 0xD7;
 
 /// SPI mode
 pub const MODE: Mode = embedded_hal::spi::MODE_3;
 
-
 /// L3GD20 driver
-pub struct L3gd20<SPI, CS> {
-    spi: SPI,
-    cs: CS,
+pub struct L3gd20<Spi> {
+    spi: Spi,
 }
 
-impl<SPI, CS, E> L3gd20<SPI, CS>
-where
-    SPI: Transfer<u8, Error = E> + Write<u8, Error = E>,
-    CS: OutputPin,
+impl<Spi: SpiDevice> L3gd20<Spi>
 {
     /// Creates a new driver from a SPI peripheral and a NCS pin
-    pub fn new(spi: SPI, cs: CS) -> Result<Self, E> {
-        let mut l3gd20 = L3gd20 { spi, cs };
+    pub fn new(spi: Spi) -> Result<Self, Spi::Error> {
+        let mut l3gd20 = L3gd20 { spi };
 
         // power up and enable all the axes
-        l3gd20.write_register(Register::CTRL_REG1, 0b00_00_1_111)?;
+        l3gd20.write_register(Register::CTRL_REG1, 0b0000_1111)?;
 
         Ok(l3gd20)
     }
 
     /// Temperature measurement + gyroscope measurements
-    pub fn all(&mut self) -> Result<Measurements, E> {
+    pub fn all(&mut self) -> Result<Measurements, Spi::Error> {
         let mut bytes = [0u8; 9];
         self.read_many(Register::OUT_TEMP, &mut bytes)?;
 
@@ -56,12 +56,12 @@ where
                 y: (bytes[5] as u16 + ((bytes[6] as u16) << 8)) as i16,
                 z: (bytes[7] as u16 + ((bytes[8] as u16) << 8)) as i16,
             },
-            temp: bytes[1] as i8,
+            temp_raw: bytes[1] as i8,
         })
     }
 
     /// Gyroscope measurements
-    pub fn gyro(&mut self) -> Result<I16x3, E> {
+    pub fn gyro(&mut self) -> Result<I16x3, Spi::Error> {
         let mut bytes = [0u8; 7];
         self.read_many(Register::OUT_X_L, &mut bytes)?;
 
@@ -72,36 +72,41 @@ where
         })
     }
 
-    /// Temperature sensor measurement
-    pub fn temp(&mut self) -> Result<i8, E> {
+    /// Raw temperature sensor measurement
+    pub fn temp_raw(&mut self) -> Result<i8, Spi::Error> {
         Ok(self.read_register(Register::OUT_TEMP)? as i8)
     }
 
+    /// Actual temperature derived by subtracting the raw measurement to the baseline value of 25 C
+    pub fn temp_celcius(&mut self) -> Result<i16, Spi::Error> {
+        Ok(25 - self.temp_raw()? as i16)
+    }
+
     /// Reads the WHO_AM_I register; should return `0xD4`
-    pub fn who_am_i(&mut self) -> Result<u8, E> {
+    pub fn who_am_i(&mut self) -> Result<u8, Spi::Error> {
         self.read_register(Register::WHO_AM_I)
     }
 
     /// Read `STATUS_REG` of sensor
-    pub fn status(&mut self) -> Result<Status, E> {
+    pub fn status(&mut self) -> Result<Status, Spi::Error> {
         let sts = self.read_register(Register::STATUS_REG)?;
         Ok(Status::from_u8(sts))
     }
 
     /// Get the current Output Data Rate
-    pub fn odr(&mut self) -> Result<Odr, E> {
+    pub fn odr(&mut self) -> Result<Odr, Spi::Error> {
         // Read control register
         let reg1 = self.read_register(Register::CTRL_REG1)?;
         Ok(Odr::from_u8(reg1))
     }
 
     /// Set the Output Data Rate
-    pub fn set_odr(&mut self, odr: Odr) -> Result<&mut Self, E> {
+    pub fn set_odr(&mut self, odr: Odr) -> Result<&mut Self, Spi::Error> {
         self.change_config(Register::CTRL_REG1, odr)
     }
 
     /// Get current Bandwidth
-    pub fn bandwidth(&mut self) -> Result<Bandwidth, E> {
+    pub fn bandwidth(&mut self) -> Result<Bandwidth, Spi::Error> {
         let reg1 = self.read_register(Register::CTRL_REG1)?;
         Ok(Bandwidth::from_u8(reg1))
     }
@@ -109,14 +114,14 @@ where
     /// Set low-pass cut-off frequency (i.e. bandwidth)
     ///
     /// See `Bandwidth` for further explanation
-    pub fn set_bandwidth(&mut self, bw: Bandwidth) -> Result<&mut Self, E> {
+    pub fn set_bandwidth(&mut self, bw: Bandwidth) -> Result<&mut Self, Spi::Error> {
         self.change_config(Register::CTRL_REG1, bw)
     }
 
     /// Get the current Full Scale Selection
     ///
     /// This is the sensitivity of the sensor, see `Scale` for more information
-    pub fn scale(&mut self) -> Result<Scale, E> {
+    pub fn scale(&mut self) -> Result<Scale, Spi::Error> {
         let scl = self.read_register(Register::CTRL_REG4)?;
         Ok(Scale::from_u8(scl))
     }
@@ -125,43 +130,29 @@ where
     ///
     /// This sets the sensitivity of the sensor, see `Scale` for more
     /// information
-    pub fn set_scale(&mut self, scale: Scale) -> Result<&mut Self, E> {
+    pub fn set_scale(&mut self, scale: Scale) -> Result<&mut Self, Spi::Error> {
         self.change_config(Register::CTRL_REG4, scale)
     }
 
-    fn read_register(&mut self, reg: Register) -> Result<u8, E> {
-        let _ = self.cs.set_low();
-
+    fn read_register(&mut self, reg: Register) -> Result<u8, Spi::Error> {
         let mut buffer = [reg.addr() | SINGLE | READ, 0];
-        self.spi.transfer(&mut buffer)?;
-
-        let _ = self.cs.set_high();
+        self.spi.transfer_in_place(&mut buffer)?;
 
         Ok(buffer[1])
     }
 
     /// Read multiple bytes starting from the `start_reg` register.
     /// This function will attempt to fill the provided buffer.
-    fn read_many(&mut self,
-                 start_reg: Register,
-                 buffer: &mut [u8])
-                 -> Result<(), E> {
-        let _ = self.cs.set_low();
-        buffer[0] = start_reg.addr() | MULTI | READ ;
-        self.spi.transfer(buffer)?;
-        let _ = self.cs.set_high();
+    fn read_many(&mut self, start_reg: Register, buffer: &mut [u8]) -> Result<(), Spi::Error> {
+        buffer[0] = start_reg.addr() | MULTI | READ;
+        self.spi.transfer_in_place(buffer)?;
 
         Ok(())
     }
 
-
-    fn write_register(&mut self, reg: Register, byte: u8) -> Result<(), E> {
-        let _ = self.cs.set_low();
-
+    fn write_register(&mut self, reg: Register, byte: u8) -> Result<(), Spi::Error> {
         let buffer = [reg.addr() | SINGLE | WRITE, byte];
         self.spi.write(&buffer)?;
-
-        let _ = self.cs.set_high();
 
         Ok(())
     }
@@ -172,7 +163,7 @@ where
     /// affecting other parts of the register that might contain desired
     /// configuration. This allows the `L3gd20` struct to be used like
     /// a builder interface when configuring specific parameters.
-    fn change_config<B: BitValue>(&mut self, reg: Register, bits: B) -> Result<&mut Self, E> {
+    fn change_config<B: BitValue>(&mut self, reg: Register, bits: B) -> Result<&mut Self, Spi::Error> {
         // Create bit mask from width and shift of value
         let mask = B::mask() << B::shift();
         // Extract the value as u8
@@ -207,7 +198,9 @@ trait BitValue {
 
 #[allow(dead_code)]
 #[allow(non_camel_case_types)]
+#[allow(clippy::upper_case_acronyms)]
 #[derive(Clone, Copy)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
 enum Register {
     WHO_AM_I = 0x0F,
     CTRL_REG1 = 0x20,
@@ -237,10 +230,9 @@ enum Register {
     INT1_DURATION = 0x38,
 }
 
-
-
 /// Output Data Rate
 #[derive(Debug, Clone, Copy)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
 pub enum Odr {
     /// 95 Hz data rate
     Hz95 = 0x00,
@@ -279,6 +271,7 @@ impl Odr {
 
 /// Full scale selection
 #[derive(Debug, Clone, Copy)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
 pub enum Scale {
     /// 250 Degrees Per Second
     Dps250 = 0x00,
@@ -321,6 +314,7 @@ impl Scale {
 /// filter. The cut-off depends on the `Odr` of the sensor, for specific
 /// information consult the data sheet.
 #[derive(Debug, Clone, Copy)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
 pub enum Bandwidth {
     /// Lowest possible cut-off for any `Odr` configuration
     Low = 0x00,
@@ -387,7 +381,8 @@ impl Scale {
 }
 
 /// XYZ triple
-#[derive(Debug)]
+#[derive(Debug, Clone, Copy)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
 pub struct I16x3 {
     /// X component
     pub x: i16,
@@ -398,16 +393,25 @@ pub struct I16x3 {
 }
 
 /// Several measurements
-#[derive(Debug)]
+#[derive(Debug, Clone, Copy)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
 pub struct Measurements {
     /// Gyroscope measurements
     pub gyro: I16x3,
-    /// Temperature sensor measurement
-    pub temp: i8,
+    /// Raw temperature sensor measurement
+    pub temp_raw: i8,
+}
+
+impl Measurements {
+    /// Convert the raw temperature value to degrees celcius
+    pub fn temp_celcius(&self) -> i16 {
+        25 - self.temp_raw as i16
+    }
 }
 
 /// Sensor status
 #[derive(Debug, Clone, Copy)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
 pub struct Status {
     /// Overrun (data has overwritten previously unread data)
     /// has occurred on at least one axis
@@ -431,14 +435,14 @@ pub struct Status {
 impl Status {
     fn from_u8(from: u8) -> Self {
         Status {
-            overrun: (from & 1 << 7) != 0,
-            z_overrun: (from & 1 << 6) != 0,
-            y_overrun: (from & 1 << 5) != 0,
-            x_overrun: (from & 1 << 4) != 0,
-            new_data: (from & 1 << 3) != 0,
-            z_new: (from & 1 << 2) != 0,
-            y_new: (from & 1 << 1) != 0,
-            x_new: (from & 1 << 0) != 0,
+            overrun: (from & (1 << 7)) != 0,
+            z_overrun: (from & (1 << 6)) != 0,
+            y_overrun: (from & (1 << 5)) != 0,
+            x_overrun: (from & (1 << 4)) != 0,
+            new_data: (from & (1 << 3)) != 0,
+            z_new: (from & (1 << 2)) != 0,
+            y_new: (from & (1 << 1)) != 0,
+            x_new: (from & 1) != 0,
         }
     }
 }